requirements.txt
fastapi
uvicorn[standard]
python-multipart
SQLAlchemy
python-jose
pydantic[email]
passlib[bcrypt]
pytz
psycopg2-binary
fastapi-pagination
openpyxl
requests
alembic
python-dotenv
python-telegram-bot